+++ src/Ampersand/Options/DaemonParser.hs
@@ -22,13 +22,13 @@ daemonOptsParser =
   <*> strOption
         ( long "daemonconfig"
         <> metavar "CONFIGFILE"
-        <> value (".ampersand")
+        <> value ".ampersand"
         <> showDefault
-        <> help ("The config file contains the list of files to be monitored.")
+        <> help "The config file contains the list of files to be monitored."
         )
   <*> fSpecGenOptsParser True
   <*> boolFlags True "warnings"
-        ( "show warnings in the output, if any. "
-        ) mempty
+         "show warnings in the output, if any. "
+         mempty