+++ ReleaseNotes.md
@@ -1,7 +1,7 @@
 ﻿# Release notes of Ampersand
 
-## Unreleased
-
+## Unreleased changes
+* [Issue #1070](https://github.com/AmpersandTarski/Ampersand/issues/1070) Remove CLI options dbHost, dbName, dbUser and dbPass. These are part of the prototype framework and can be specified as config or environment variable as described here: https://github.com/AmpersandTarski/prototype/tree/development/config 
 * [Issue #1089](https://github.com/AmpersandTarski/Ampersand/issues/1089) Add template attributes to VIEW syntax
 
 ## v4.1.2 (9 october 2020)